Implement element attributes for Particles, survival analysis, hand tracking.
In order to have Particles objects with elements from several sources,
one needs to handle such sources and parameters. Because Particle
objects are spawned as views of DataFrames, and that DataFrames are not
very subclassing-friendly, the simplest way to do it was to inherit said
attributes from the group (Particles)

Particles maintain the attributes for each underlying Particle in an
"element attribute", a dict with particle_number: attribute_value
pairing. When a Particle is spawned, it inherits the group Particles and
will fetch its attributes from it, as long as said attributes names are
listed in the Particle._element_attributes set.

Likewise, when setting an attribute in Particle, this will look whether
it's an element attribute, and if so, will let the group handle the
writing of the attribute.

Other work:
- Can now save DNA molecules without the source plugged in
- Fixed the module for hand tracking from kymograms
- Implemented survival analysis tools
- Filtering and tracking parameters are written in the Particles object
as element attributes.