Fix extension loading and Particles attribute errors.
I had forgotten to reinstate the inheritance of Particle attributes from
Particles.

In addition, the loading of extensions was broken because
extensions should not import cats modules with extendable classes at
loading/import time. This causes the extendable classes to load and be
decorated before all of the extensions are loaded, and hence it prevents
the full loading of extensions.